refactored document parsing
