+++ create_cr_and_diff.py
@@ -3,8 +3,8 @@ import os
 from dataclasses import asdict
 
 from difftool.diffmaker import CRDiffMaker
-from parsing.cr import extract_cr
-from parsing.utils.formatter import CRFormatterFactory
+from extractor.cr import extract_cr
+from extractor.formatter import CRFormatterFactory
 
 
 def diff(old_txt, new_txt, old_set_code=None, new_set_code=None, forced_matches=None):
