+++ src/scraper/cr_scraper.py
